uws: cast the trailer length to int like the sibling Super::write calls
AsyncSocket::write takes an int length and every other dynamic-length
Super::write in HttpResponse.h already casts (int); make the trailer write
match so it does not trip -Wshorten-64-to-32.

+++ packages/bun-uws/src/HttpResponse.h
@@ -147,7 +147,7 @@ public:
                 Super::write("0\r\n\r\n", 5);
             } else {
                 Super::write("0\r\n", 3);
-                Super::write(trailer.data(), trailer.length());
+                Super::write(trailer.data(), (int) trailer.length());
                 Super::write("\r\n", 2);
             }
             httpResponseData->markDone(this);
